Issue SB-24793 feat: Review comments resolved
